refactor: 패키지 구조에 맞게 엔티티 경로 수정
